+++ kernel/sync/mqueue.h
@@ -14,8 +14,8 @@ struct mqueue {
 
 void kernel_mqueue_init(struct mqueue *mqueue, void *buf, uint16_t length, uint16_t element_size);
 void kernel_mqueue_peek(struct mqueue *mqueue, void **data);
-enum bw_status kernel_mqueue_send(struct mqueue *mqueue, void *data, uint32_t timeout_ms);
-enum bw_status kernel_mqueue_receive(struct mqueue *mqueue, void *data, uint32_t timeout_ms);
+enum bw_status kernel_mqueue_send(struct mqueue *mqueue, void *data, uint64_t timeout_ms);
+enum bw_status kernel_mqueue_receive(struct mqueue *mqueue, void *data, uint64_t timeout_ms);
 void kernel_mqueue_overwrite(struct mqueue *mqueue, void *data);
 
 #endif
